requirements.txt
# Core dependencies for Crypto Analysis Pro Dashboard
streamlit>=1.31.0
agno>=0.1.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
requests>=2.31.0
google-cloud-core>=2.3.0
google-generativeai>=0.3.0

# Visualization enhancements
matplotlib>=3.7.0
seaborn>=0.13.0

# Formatting and development tools
python-dotenv>=1.0.0
black>=23.12.0
flake8>=6.1.0
pytest>=7.4.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
scikit-learn>=1.3.0